+++ example/http/main.cpp
@@ -75,6 +75,9 @@ int main() {
 	}
 	std::cout << std::endl;
 
+	std::this_thread::sleep_for(std::chrono::seconds(2));
+	server.stop();
+
 	w.join();
 
 	return 0;
