+++ include/http/fd.hpp
@@ -2,6 +2,7 @@
 #pragma once
 
 #include <utility>
+#include <functional>
 
 #include <unistd.h>
 
@@ -9,42 +10,23 @@ namespace http {
 
 struct fd_t {
 	int fd;
-	
-	fd_t(): 
-		fd(-1) {
-	}
 
-	fd_t(int new_fd):
-		fd(new_fd) {
-	}
+	fd_t();
+
+	fd_t(int new_fd);
 
 	fd_t(const fd_t& other) = delete;
 	fd_t& operator=(const fd_t& other) = delete;
 
-	fd_t(fd_t&& other):
-		fd(std::exchange(other.fd, -1)) {
-	}
+	fd_t(fd_t&& other);
 
-	fd_t& operator=(fd_t&& other) {
-		close();
-		fd = std::exchange(other.fd, -1);
-		return *this;
-	}
+	fd_t& operator=(fd_t&& other);
 
-	~fd_t() {
-		close();
-	}
+	~fd_t();
 
-	operator bool() const {
-		return fd != -1;
-	}
+	operator bool() const;
 
-	void close() {
-		if (fd != -1) {
-			::close(fd);
-			fd = -1;
-		}
-	}
+	void close();
 };
 
 }
