--- src/http/tmp.cpp
@@ -1,2 +1,0 @@
-
-void tmp() {}
